requirements.txt
Flask==1.1.1
pytesseract==0.3.13
Pillow==8.0.0
ollama
